fix: Fix CPU ffmpeg images.
