+++ scripts/make/attach.mk
@@ -1,0 +1,3 @@
+.PHONY: attach
+attach:		## Attach to the docker container
+	docker attach $(NAME)
