+++ scripts/make/stop.mk
@@ -1,0 +1,4 @@
+.PHONY: stop
+stop:		## Stops docker container
+	-docker stop $(NAME)
+	-docker rm $(NAME)
