+++ tests/cpu/test_imports.py
@@ -1,0 +1,22 @@
+class TestImports:
+
+    def test_imports(self):
+        import numpy
+        import cv2
+        import sympy
+        import scipy
+        import matplotlib
+        import pandas
+        import sklearn
+        import skimage
+        import PIL
+        import albumentations
+        import zmq
+        import Cython
+        import requests
+        import psutil
+        import pydantic
+        import yaml
+        import notebook
+        import ipywidgets
+        import tqdm
