+++ tests/gpu/test_torchvision.py
