--- tests/test_imports.py
@@ -1,17 +1,0 @@
-class TestImports:
-
-    def test_imports(self):
-        import numpy
-        import cv2
-
-        import tensorrt
-        import torch
-        import torchvision
-        import torchaudio
-
-        import argus
-        import kornia
-        import timm
-
-        import PyNvCodec
-        import PytorchNvCodec
